perf_event: added 'metric' field to pe_event_info_t that is set for TOPDOWN events and ensures reset
when rdpmc is enabled, if topdown events are allowed to accumulate they 
produce innacurate results. Therefore, a new field needs to be added
for topdown events that ensures their reset flag is not disabled.

The field is named 'metric' instead of 'topdown' so that in case events 
are developed in the future with similar behavior to topdown events, 
the name will still make sense.