topdown: simplified metrics calculation
Previously, the topdown component calculated metrics by taking the
difference of the metrics before and the metrics after the calculated
code block using Equation 1:

M% = (Mb*Sb/255 - Ma*Sa/255) / (Sb - Sa) * 100                   (1)

where Mx are the raw bytes of the metric before and after the calipered
code block and Sx are the slots. However, if Sa = 0 this simplifies to

M% = Mb/255 * 100                                                (2)

Therefore it is sufficient to simply reset the PERF_METRICS MSR and
SLOTS during PAPI_start() and then use Equation 2 in PAPI_stop, reducing
the number of dangerous rdpmc calls, reducing overhead, and simplifying
the code.